refact: 일렉트론 SOLID하게 책임 분리
